HTMBackend/requirements.txt
APScheduler==3.6.3
asteval==0.9.1
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configobj==5.0.6
coverage==3.7.1
cycler==0.10.0
# DBUtils
ebaysdk==2.2.0
execnet==1.9.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==4.3.2
funcsigs
futures==3.3.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonlib==1.6.1
kiwisolver==1.1.0
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==2.2.5
mock==1.0.1
mysql-connector==2.2.9
numpy==1.12.1
nupic==1.0.5
nupic.bindings==1.0.6
ordereddict==1.1
pandas==0.24.2
prettytable==0.7.2
psutil==1.0.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.6.2
pyparsing==2.4.7
pyproj==1.9.3
pytest==3.0.7
pytest-cov==2.5.0
pytest-xdist==1.16.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-engineio==3.14.2
python-gnupg==0.4.7
pytz==2021.1
PyYAML==3.10
requests==2.25.1
schedule==0.6.0
simplejson==3.17.0
six==1.16.0
soupsieve==1.9.6
trollius==2.2
typing==3.10.0.0
tzlocal==2.0.0
unittest2==0.5.1
urllib3==1.26.6
validictory==0.9.1
Werkzeug==1.0.1
wincertstore==0.2
